tools(mlflow): unify extractors into one --source CLI
Merge extract_interactive_events.py into extract_permission_events.py
behind a --source {headless,interactive} flag. The split was never
bash-vs-other: both readers handle all tool types; it was headless
(MLflow, MCP-only) vs interactive (transcripts, Bash exposed). Now:

- shared classifiers / followup / writers / event schema in one module
- two source readers (collect_headless / collect_interactive) dispatched
  by --source, emitting the same schema with a 'source' column
- source-specific summaries (headless denial stats / interactive outcome
  + bash + approval-candidate breakdown)

Removes the duplicated helpers across the two former scripts. Tests for
both sources merged into test_extract_permission_events.py.
